Remove support for really old versions of salt
It seems this conditional was added to make use of some new salt
functionality, but it was implemented badly and would only work with a
specific version of salt. We don't care about really old versions of
salt so it should be safe to remove the conditional and associated code
for old salt versions.