Default to using pools for Debian 9
The upstream ntp.conf uses `pool`s instead of `server`s, so lets
replicate that as the default behaviour.

Apparently the difference is that with `server` the DNS lookup is done
once and cached, while with `pool` the DNS is regularly re-queried up to
pick up other IPs in the pool that the entry may resolve to.